Небольшие изменения.
