+++ app/src/main/java/ru/geekbrains/bookofrecipes/presentation/ui/details/DetailsFragment.kt
@@ -1,15 +1,10 @@
 package ru.geekbrains.bookofrecipes.presentation.ui.details
 
-import android.annotation.SuppressLint
-import android.content.Context
 import android.graphics.Color
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.annotation.AttrRes
-import androidx.annotation.ColorInt
-import androidx.core.content.res.use
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
 import com.google.android.material.transition.MaterialContainerTransform
@@ -17,23 +12,13 @@ import kotlinx.android.synthetic.main.detail_fragment.view.*
 import ru.geekbrains.bookofrecipes.R
 import ru.geekbrains.bookofrecipes.presentation.models.RecipeModelForRecycler
 import ru.geekbrains.bookofrecipes.service.extensions.loadFromUrl
-
-@ColorInt
-@SuppressLint("Recycle")
-fun Context.themeColor(
-        @AttrRes themeAttrId: Int
-): Int {
-    return obtainStyledAttributes(
-            intArrayOf(themeAttrId)
-    ).use {
-        it.getColor(0, Color.MAGENTA)
-    }
-}
+import ru.geekbrains.bookofrecipes.service.extensions.themeColor
 
 class DetailsFragment : Fragment() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+
         sharedElementEnterTransition = MaterialContainerTransform().apply {
             drawingViewId = R.id.nav_host_fragment
             duration = resources.getInteger(R.integer.reply_motion_duration_small).toLong()
@@ -43,8 +28,8 @@ class DetailsFragment : Fragment() {
     }
 
     override fun onCreateView(
-            inflater: LayoutInflater, container: ViewGroup?,
-            savedInstanceState: Bundle?
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
     ): View? {
         return inflater.inflate(R.layout.detail_fragment, container, false)
     }
@@ -54,9 +39,10 @@ class DetailsFragment : Fragment() {
 
         view.navigation_icon.setOnClickListener {
             findNavController().navigateUp()
-
         }
+
         val recipeData = arguments?.getSerializable("recipe") as RecipeModelForRecycler
+
         view.title_detail_dish.text = recipeData.title
         view.full_detail_dish.text = recipeData.summary
         view.detail_image_dish.loadFromUrl(recipeData.imageUrl)
