+++ app/src/main/java/ru/geekbrains/bookofrecipes/service/extensions/Context.kt
@@ -1,0 +1,20 @@
+package ru.geekbrains.bookofrecipes.service.extensions
+
+import android.annotation.SuppressLint
+import android.content.Context
+import android.graphics.Color
+import androidx.annotation.AttrRes
+import androidx.annotation.ColorInt
+import androidx.core.content.res.use
+
+@ColorInt
+@SuppressLint("Recycle")
+fun Context.themeColor(
+    @AttrRes themeAttrId: Int
+): Int {
+    return obtainStyledAttributes(
+        intArrayOf(themeAttrId)
+    ).use {
+        it.getColor(0, Color.MAGENTA)
+    }
+}
